# schemas/orchestration/api.openapi.yaml
openapi: 3.1.0
info:
  title: CodeBridge Orchestration API
  version: 1.0.0-draft
  description: |
    Session-first API for Agent sessions, folders, flows, runs, approvals and discovery.
    TaskRecord/WorkItem routes remain available as an internal compatibility surface.
  x-codebridge-status: session-first-target
servers:
  - url: /
paths:
...
  /v1/sessions/{session_id}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      operationId: getSession
      summary: Get a session and its current context
      responses:
        '200':
          description: Session projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSession'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/flows/{flow_id}:
    parameters:
      - $ref: '#/components/parameters/FlowId'
    get:
      operationId: getFlow
      summary: Get a Flow definition and revision
      responses:
        '200':
          description: Flow definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/flows/{flow_id}/review:
    parameters:
      - $ref: '#/components/parameters/FlowId'
    post:
      operationId: reviewFlow
      summary: Approve or reject a Workflow Candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                git_revision:
                  type: [string, 'null']
                  description: Required for approval; points to the reviewed Git definition.
      responses:
        '200':
          description: Updated Flow review state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/flows/{flow_id}/apply:
    parameters:
      - $ref: '#/components/parameters/FlowId'
    post:
      operationId: applyFlow
      summary: Apply a Flow to the next Run of a Session
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id: { type: string }
      responses:
        '200':
          description: Flow binding recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Acknowledgement'
  /v1/flows/candidates:
    post:
      operationId: saveFlowCandidate
      summary: Save a Session-generated Flow as a Candidate
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowCandidateRequest'
      responses:
        '201':
          description: Flow Candidate saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
...
components:
  parameters:
...
    SessionId:
      in: path
      name: session_id
      required: true
      schema:
        type: string
    FlowId:
      in: path
      name: flow_id
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: State or idempotency conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    AgentSession:
      type: object
      required: [session_id, agent_id, provider_session_id, folder_id, flow_id, status]
      properties:
        session_id: { type: string }
        agent_id: { type: string }
        provider_session_id: { type: string }
        task_record_id: { type: [string, 'null'] }
        folder_id: { type: [string, 'null'] }
        flow_id: { type: [string, 'null'] }
        model: { type: [string, 'null'] }
        flow_revision: { type: [string, 'null'] }
        cwd: { type: [string, 'null'] }
        additional_directories:
          type: array
          items: { type: string }
        title: { type: [string, 'null'] }
        status:
          type: string
          enum: [active, idle, closed, unavailable]
        last_seen: { type: [string, 'null'], format: date-time }
...
    Flow:
      type: object
      required: [flow_id, kind, status, source, definition_revision, steps]
      properties:
        flow_id: { type: string }
        name: { type: [string, 'null'] }
        kind:
          type: string
          enum: [ephemeral, guide, runbook]
        status:
          type: string
          enum: [draft, candidate, published, deprecated]
        source:
          type: string
          enum: [agent_generated, user_selected, git]
        definition_revision: { type: string }
        review_status:
          type: string
          enum: [pending, approved, rejected]
        git_revision: { type: [string, 'null'] }
        validation_issues:
          type: array
          items: { type: string }
        steps:
          type: array
          items:
            type: object
            required: [id]
            properties:
              id: { type: string }
              capability: { type: [string, 'null'] }
              purpose: { type: [string, 'null'] }
              depends_on:
                type: array
                items: { type: string }
              branches:
                type: array
                items:
                  type: object
                  required: [when, next]
                  properties:
                    when: { type: string }
                    next: { type: string }
        session_id: { type: [string, 'null'] }
    FlowCandidateRequest:
      type: object
      required: [session_id, definition_revision]
      properties:
        session_id: { type: string }
        run_id: { type: [string, 'null'] }
        definition_revision: { type: string }
        flow:
          type: object
          additionalProperties: true
...
    Acknowledgement:
      type: object
      required: [request_id, accepted]
      properties:
        request_id:
          type: string
        accepted:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
            request_id:
              type: string